refactor: 피드 조회 filter 기능 리팩토링
