+++ backend/smody/src/main/java/com/woowacourse/smody/challenge/service/ChallengeQueryService.java
@@ -11,7 +11,7 @@ import com.woowacourse.smody.challenge.dto.ChallengeOfMineResponse;
 import com.woowacourse.smody.challenge.dto.ChallengeResponse;
 import com.woowacourse.smody.challenge.dto.ChallengeTabResponse;
 import com.woowacourse.smody.challenge.dto.ChallengersResponse;
-import com.woowacourse.smody.common.PagingParams;
+import com.woowacourse.smody.db_support.PagingParams;
 import com.woowacourse.smody.cycle.domain.Cycle;
 import com.woowacourse.smody.cycle.service.CycleService;
 import com.woowacourse.smody.exception.BusinessException;
