+++ backend/smody/src/main/java/com/woowacourse/smody/cycle/repository/DynamicCycleRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.woowacourse.smody.cycle.repository;
 
-import com.woowacourse.smody.common.PagingParams;
+import com.woowacourse.smody.db_support.PagingParams;
 import com.woowacourse.smody.cycle.domain.Cycle;
 import java.time.LocalDateTime;
 import java.util.List;
