+++ backend/smody/src/main/java/com/woowacourse/smody/feed/controller/FeedController.java
@@ -1,6 +1,6 @@
 package com.woowacourse.smody.feed.controller;
 
-import com.woowacourse.smody.common.PagingParams;
+import com.woowacourse.smody.db_support.PagingParams;
 import com.woowacourse.smody.feed.dto.FeedResponse;
 import com.woowacourse.smody.feed.service.FeedQueryService;
 import java.util.List;
