+++ backend/smody/src/main/java/com/woowacourse/smody/feed/repository/FeedDynamicRepository.java
@@ -1,6 +1,6 @@
 package com.woowacourse.smody.feed.repository;
 
-import com.woowacourse.smody.common.PagingParams;
+import com.woowacourse.smody.db_support.PagingParams;
 import com.woowacourse.smody.feed.domain.Feed;
 import java.util.List;
